+++ crates/dprint/src/utils/glob/mod.rs
@@ -1,0 +1,7 @@
+mod glob;
+mod glob_matcher;
+mod glob_utils;
+
+pub use glob::glob;
+pub use glob_matcher::*;
+pub use glob_utils::*;
