+++ crates/dprint/src/utils/mod.rs
@@ -4,7 +4,7 @@ mod file_path_utils;
 mod file_text;
 mod get_bytes_hash;
 mod get_difference;
-mod glob_utils;
+mod glob;
 mod path_source;
 mod pretty_print_json_text;
 mod reset_events;
@@ -19,7 +19,7 @@ pub use file_path_utils::*;
 pub use file_text::*;
 pub use get_bytes_hash::*;
 pub use get_difference::*;
-pub use glob_utils::*;
+pub use glob::*;
 pub use path_source::*;
 pub use pretty_print_json_text::*;
 pub use reset_events::*;
